fix course_id to courseOverview fragment

+++ app/src/main/java/com/projectse/aads/task_tracker/MainActivity.java
@@ -180,7 +180,7 @@ public class MainActivity
 
 
     @Override
-    public void callCourseOverviewFragment(int course_id) {
+    public void callCourseOverviewFragment(long course_id) {
         CourseOverviewFragment fragment = new CourseOverviewFragment();
         setCurrentFragment(fragment);
         fragment.setCourseID(course_id);
